% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDs.R
\name{getIDs}
\alias{getIDs}
\title{get the gene symbols of probes for each GPL}
\usage{
getIDs(gpl)
}
\arguments{
\item{gpl}{the name of GPL, such as GPL570}
}
\value{
a data.frame which contains the gene symbols of probes
}
\description{
getIDs need the name of GPL, such as GPL570, and it will return a data.frame
}
\examples{
getIDs('gpl570')
}
